Add --enable-shared option to iosbuild.sh to build dynamic framework
Enabling the shared framework bumps the iOS deployment target to 8.0,
the minimum required to support dynamic framework deployment in apps.

When not using --enable-shared, a static library for iOS 6.0+ will still
be built.

Minimum version settings have been moved into ios-version.sh so they
can be updated in a single place.

Note that due to usage of inttypes.h which clang does not define as
a modular header, users of the resulting VPX.framework will need to
set the CLANG_ALLOW_NON_MODULAR_INCLUDES_IN_FRAMEWORK_MODULES setting
in their build configuration.

A module map is thus not included, since it can't be enforced.

BUG=https://bugs.chromium.org/p/webm/issues/detail?id=1092
Change-Id: I28fb06ce65e48ed167a88c14a7bfb2861989317e