Emscripten multithreading hack for VP8 decoding
Busywaits are very inefficient in emscripten because there's no
way for a threading to yield to the scheduler in the middle of
a code run.

This lead to very bad performance for multithreaded VP8 decoder
under system load, because threads would spin waiting on each other
very unpleasantly.

Uses a mutex and condition variable around the macroblocks, which
lets the thread that's waiting actually wait. Seems to behave
better under load than the current version, at least under
Safari Technical Preview.