Microoptimize VP8 loop filter a little for emscripten
Use full-size ints instead of chars since we get no auto
vectorization magic.

Add a lookup table for clamping instead of branching, since
there's no native instruction in asm.js or wasm for clamps
and the branches seem more expensive.

This gives a noticeable 5-10% speedup in decode times for VP8
on Edge 16 and Safari 11 (wasm), with more modest
improvements for IE 11 and Safari and Edge (asm.js)

This is pretty hacky, and should be instead done similarly
to how asm-optimized versions are done for other platforms.